requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
attrs==23.1.0
backoff==2.2.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cohere==4.35
fastavro==1.8.2
Flask==3.0.0
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
